allow the full range of session handles
In the current code, session handles are only allowed in the range
02000000-0200003f; 03000000-0300003f with TPM_RC_VALUE being given if a
value is unmarshalled outside that range.  This rather conflicts with
the embryonic Linux in-kernel Resource Manager because it virtualizes
handles by counting down from the top of the handle range (to be as
different as possible from physical handles).  This means that the TSS
needs to allow the handle range to be 02000000-02ffffff; 03000000
-03ffffff.  Fix that by changing the defines in TPM_Types.h

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>