Support building the TSS library without using files for state
This patch defines --enable-nofile to build the TSS library without
using files for storing temporary or persistent state.  All state is
stored in the TSS context and is lost when the context is deleted.

Refer to the ibmtss documentation, "Build Options" section on
"-DTPM_TSS_NOFILE".

(Fails to build properly with TPM 1.2 enabled.)

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>
Reviewed-by: Stefan Berger <stefanb@linux.ibm.com>
Reviewed-by: George Wilson <gcwilson@linux.ibm.com>
Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>