Install TPM 1.2 and TPM 2.0 utilities with different names
When installing both TPM 1.2 and TPM 2.0 utilities, one version
overwrites the other version.  Although the "--program-prefix=" and
"--program-suffix=" ./configure options can be used to rename programs
on "make install", the defined prefix and suffix are for the entire
package.

This patch hard codes the prefix for utilities in the utils/ and
utils12/ directories (eg. tsspcrread, tss1pcrread).

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>
Reviewed-by: Stefan Berger <stefanb@linux.ibm.com>
Reviewed-by: George Wilson <gcwilson@linux.ibm.com>

Changelog:
- Changed from TPM 1.2 utility program names, changing the prefix to "tss1"
and removing the suffix.
Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>