utils: Add a tool to generate the public name of a key
This tool is heavily based on loadexternal, and basically takes the
same parameters.  In our build system we need to generate a few of the
magic files for use with policy authorise.  To do this we need to know
the public name of the keys, but the build system does not have a TPM
that we can use.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>