TSS: Further separate the TPM 1.2 and TPM 2.0 dependencies.
Add ifdefs for 1.2 and 2.0 to header files so a 2.0 build does not
pull in unnecessary structures and API declarations.  Add the 1.2
structures to the parameter unions, which are used to determine the
maximum size of the unions.

Change the 2.0 eventextend PCR index type, because it no longer sees
the 1.2 typedef.

Add netinet to pcrread and policysigned, since they no longer see
the 1.2 header.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>