TSS: Add crypto library independent functions to tsscrypto.c
Add TSS_RsaNew(), TSS_RsaFree(), and TSS_RSAGeneratePublicTokenI()
that use void * so they are not openssl dependent.  Deprecate the
TSS_RSAGeneratePublicTokenI() function that used the openssl
structure.

Move the openssl CURVE_DATA structure to the .c file, comment and
clean up the CURVE_DATA memory free code, which was introducing a
memory leak on openssl in case of an error.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>