UTILS: Make ekutils library independent
This change is backward compatible (even though it's just sample code).

In the header, library independent functions use a void * rather than
the openssl structure.  Added ifdef macros for openssl and Mark
functions that are openssl specific (that use those structures and
that are not intended general use).

Update the source to use the void * to match the header.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>