UTILS: Introduce getcryptolibrary utility
This utility returns the crypto library that the utilities were
compiled with. targets are added to makefiles.

The intended use is any regression test that depends on a function
that is not supported yet in an implementation using a specific crypto
library.  The regression test can temporarily skip that test rather
than return an expected error.

For example, ECC salted session and X509 certificate creation tests
are skipped for the mbedtls implementation.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>