TSS; Add support for TPM2_NV_Certify with large data size.
TPM 2.0 rev 155 permits certifying NV data larger than that can be put
in the attestation structure.  If the size parameter is zero, the NV
data is hashed and the result put in the structure.

Add new TPM_ST_ATTEST_NV_DIGEST tag, the selector in the attestation
union, and the TPMS_NV_DIGEST_CERTIFY_INFO structure. Add unmarshal
support for the TPM_ST_ATTEST_NV_DIGEST tag and the
TPMS_NV_DIGEST_CERTIFY_INFO structure.  Add tracing support.

Modify the nvcertify utility to output the certify data for regression
testing, which is now either the data or a hash of the data.  Remove
the check for size != 0 since it is now legal.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>