REGTEST: Add a regression test for the TPM rev 155 NV certify change
Add a call from the main regression test scripts.  Add a new 155
script that iterates through the session, hash, and signature
algorithms as in the previous test, but with size 0.  Add a test that
the attestation structure contains the digest.

Remove the misplaced cleanup of msg.bin in testcredential, since it
it is needed in this test.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>