TSS/Eventlog: Adding multiple helpers for eventlib
This introduces multiple functions to help and deal with Eventlog data,
specifically for structures that belong to event data but also to other
structures, and thus it's adding similar code to deal with the little
endian used when marshaling/unmarshaling eventlog data to/from memory
buffer.

For instance. PCR extension operations don't use a specfic
endianess, but also marshall/unmarshal Algorithm ID, which is also present
at TSS_Event2 which needs to be LE when marshalling and unmarshalling

Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>
Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>