tss: Fix regression.  Need Names for no-auth policy session
This was a regression when optimizing the Name getter.  The Name is
also needed for the case of a salted policy session, even when the
policy does not include policyauthvalue, because the salt creates an
HMAC key.

Add policies to the regression test primary and RSA storage keys. Add
salt regression test.  The policy no-auth reproduces the bug.

Change initkeys.sh, testpcr.sh, testnvpin.sh and testshutdown.sh and
the Windows equivalent to use initprimary() so the primary key
template is isolated to one function.

Add tssutilsverbose.c to enable building a minimal TSS that does
not use the sample utilities.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>