travis: add initial docker/travis matrix support
Include initial docker-travis matrix to build, install, and run the
TSS regression tests on different docker images.  Running the latest
TSS regression tests in a docker-travis environment has a dependency
on the latest software TPM, which is downloaded as a tar, compiled and
installed in /usr/local/bin.

To run the TSS 1.2 regression tests would require building and
installing a separate software TPM, which supports TPM 1.2.  As a
result this patch is limited to running just the TSS 2.0 regression
tests.

This patch is unabashedly taken from the ima-evm-utils package, which
was contributed by Petr Vorel, and trimmed, removing unnecessary
packages.  Any mistakes are my own.

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>