utils: Enhance imaextend for event log types
Add -ty parameter to support pad and hash IMA logs.  Add
an option to check the template hash.  Add an option
to output the simulated PCR 10.

Add Linux and Windows regression tests for types 1 and 2, checking the
hash, and checking the simulated PCR 10 against PCR read.  Change the
regression test to loops for easier maintenance.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>