regtest: Add some items to support libtpms
Set TPM_DEVICE="/dev/tpm0", because the regression test should never
go through the resource manager.  Override any default.

Add a test for the first getrandom, so the regression test exits if it
fails.  The regression test was using the result as the session key,
which gave an odd error message when getrandom returned an error.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>