tss: Add support for TPM2_ECC_Encrypt and Decrypt
Add entries in the attribute table.  Add marshal and unmarshal
support.  Enable the KDF2 algorithm.  Enable the encrypt and decrypt
commands and add to the dispatch tables.  Add to the in and out
parameter unions.  Add debug print support.  Addd to the build makefiles.

regtest:  Add test cases.  Factor the ECC curve list.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>